Add global hub to base app component
