--- src/containers/infoHub.jsx
@@ -1,30 +1,0 @@
-class InfoHub extends React.PureComponent {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-    };
-  }
-
-  render() {
-    return null;
-  }
-}
-
-InfoHub.propTypes = {
-
-};
-
-InfoHub.defaultProps = {
-
-};
-
-const mapStateToProps = (state) => ({
-});
-
-function mapDispatchToProps(dispatch) {
-  return {
-  };
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(OraclePage);
