+++ src/scenes/App/index.js
@@ -5,11 +5,12 @@ import { Debounce } from 'react-throttle';
 import { WindowResizeListener } from 'react-window-resize-listener';
 import { withStyles } from 'material-ui/styles';
 
+import styles from './styles';
 import AppRouter from './router';
 import AppLoad from './load';
+import GlobalHub from './globalHub';
 import appActions from '../../redux/app/actions';
 import BottomBar from '../../components/BottomBar/index';
-import styles from './styles';
 import TopBar from '../../components/TopBar/index';
 
 const { toggleAll } = appActions;
@@ -30,6 +31,7 @@ export class App extends React.PureComponent {
               )}
           />
         </Debounce>
+        <GlobalHub />
         <AppLoad />
         <TopBar url={url} handler={this.props.handler} />
         <div className={classes.container}>
